requirements.txt
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.1
google-auth==2.23.3
google-auth-oauthlib==1.1.0
gspread==5.12.0
idna==3.4
oauthlib==3.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-dotenv==1.0.0
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
tiktoken==0.5.1
urllib3==2.0.7
